+++ native/oxc_ex_nif/src/lib.rs
@@ -2,6 +2,7 @@ mod bundle;
 mod codegen;
 mod error;
 mod imports;
+mod native_pipeline;
 mod options;
 mod parse;
 mod transform_many;
@@ -11,6 +12,7 @@ use rustler::{Env, NifResult, Term};
 use bundle::{bundle_entry_impl, bundle_impl, bundle_run_impl};
 use codegen::codegen_impl;
 use imports::select_impl;
+use native_pipeline::codegen_native_impl;
 use parse::{minify_impl, parse_impl, transform_impl, valid_impl};
 use transform_many::transform_many_impl;